refactor(daemon): keep readiness modules within policy
