+++ crates/ctx-daemon-service/src/daemon/tests.rs
@@ -1,3 +1,6 @@
+#[path = "semantic_activation_backoff_tests.rs"]
+mod semantic_activation_backoff;
+
 #[path = "tests/startup_recovery.rs"]
 mod startup_recovery;
 
@@ -1133,21 +1136,6 @@ fn pending_source_refresh_wait_ignores_unreachable_work_until_retry() {
     assert!(wait_for > StdDuration::from_secs(4));
     assert!(wait_for <= StdDuration::from_secs(5));
     assert!(daemon_scheduled_refresh_due(Some(&coordinator), 1_000));
-}
-
-#[test]
-fn semantic_activation_retry_wakes_the_idle_daemon_at_its_deadline() {
-    let now = Instant::now();
-    let mut runtime = DaemonRuntime::default();
-    runtime.semantic_activation_retry.consecutive_failures = 1;
-    runtime.semantic_activation_retry.retry_not_before = Some(now + StdDuration::from_secs(5));
-    runtime.semantic_activation_retry.retry_not_before_at_ms =
-        Some(utc_now().timestamp_millis() + 5_000);
-
-    let wait_for = daemon_wait_duration(&runtime, None, now + StdDuration::from_secs(30), now);
-
-    assert!(wait_for > StdDuration::from_secs(4));
-    assert!(wait_for <= StdDuration::from_secs(5));
 }
 
 fn test_daemon_run_args() -> DaemonRunArgs {
